+++ base_aux/base_objects/obj_types.py
@@ -1,5 +1,7 @@
 from typing import *
 
+from base_aux.base_source import *
+
 
 # =====================================================================================================================
 class _Cls:
@@ -40,16 +42,14 @@ class TYPES:
 
 # =====================================================================================================================
 @final
-class TypeChecker:
-
+class TypeCheck(InitSource):
     # -----------------------------------------------------------------------------------------------------------------
     @staticmethod
-    def check__name_is_build_in(name: str) -> bool:
+    def _name_is_buildin(name: str) -> bool:
         return name.startswith("__") and name.endswith("__") and len(name) > 4
 
     # -----------------------------------------------------------------------------------------------------------------
-    @staticmethod
-    def check__bool_none(source: Any) -> bool:
+    def check__bool_none(self) -> bool:
         """
         GOAL
         ----
@@ -62,20 +62,17 @@ class TypeChecker:
         ---------------------
         funcs.Valid.compare_doublesided
         """
-        return isinstance(source, (bool, type(None)))
+        return isinstance(self.SOURCE, (bool, type(None)))
 
-    @staticmethod
-    def check__elementary(source: Any) -> bool:
-        if callable(source):
+    def check__elementary(self) -> bool:
+        if callable(self.SOURCE):
             return False
-        return isinstance(source, TYPES.ELEMENTARY)
+        return isinstance(self.SOURCE, TYPES.ELEMENTARY)
 
-    @staticmethod
-    def check__elementary_single(source: Any) -> bool:
-        return isinstance(source, TYPES.ELEMENTARY_SINGLE)
+    def check__elementary_single(self) -> bool:
+        return isinstance(self.SOURCE, TYPES.ELEMENTARY_SINGLE)
 
-    @staticmethod
-    def check__elementary_single_not_none(source: Any) -> bool:
+    def check__elementary_single_not_none(self) -> bool:
         """
         its just an Idea!!!
 
@@ -88,27 +85,23 @@ class TypeChecker:
         ---------------------
         ensure_collection somewhere!
         """
-        return TypeChecker.check__elementary_single(source) and source is not None
+        return self.check__elementary_single() and self.SOURCE is not None
 
-    @staticmethod
-    def check__elementary_collection(source: Any) -> bool:
+    def check__elementary_collection(self) -> bool:
         """
         GOAL
         ----
         MOST PREFERRED to use for ensure_collection! and apply for Args!
         all other base_objects (ClsInst) will covered by brackets!
         """
-        return isinstance(source, TYPES.ELEMENTARY_COLLECTION)
+        return isinstance(self.SOURCE, TYPES.ELEMENTARY_COLLECTION)
 
-    @staticmethod
-    def check__elementary_collection_not_dict(source: Any) -> bool:
-        return isinstance(source, TYPES.ELEMENTARY_COLLECTION) and not isinstance(source, dict)
+    def check__elementary_collection_not_dict(self) -> bool:
+        return isinstance(self.SOURCE, TYPES.ELEMENTARY_COLLECTION) and not isinstance(self.SOURCE, dict)
 
     # -----------------------------------------------------------------------------------------------------------------
-    @staticmethod
     def check__iterable(
-            # self,
-            source: Any,
+            self,
             dict_as_iterable: bool = True,
             str_and_bytes_as_iterable: bool = True,
     ) -> bool:
@@ -121,30 +114,28 @@ class TypeChecker:
         :param str_and_bytes_as_iterable: usually in data processing you need to work with str-type elements as OneSolid element
             but not iterating through chars!
         """
-        if isinstance(source, dict):
+        if isinstance(self.SOURCE, dict):
             return dict_as_iterable
-        elif isinstance(source, (str, bytes)):
+        elif isinstance(self.SOURCE, (str, bytes)):
             return str_and_bytes_as_iterable
-        elif isinstance(source, (tuple, list, set,)):  # need to get it explicitly!!!
+        elif isinstance(self.SOURCE, (tuple, list, set,)):  # need to get it explicitly!!!
             return True
-        elif hasattr(source, '__iter__') or hasattr(source, '__getitem__'):
+        elif hasattr(self.SOURCE, '__iter__') or hasattr(self.SOURCE, '__getitem__'):
             return True
 
         # FINAL ---------------------
         return False
 
-    @staticmethod
-    def check__iterable_not_str(source: Any) -> bool:
+    def check__iterable_not_str(self) -> bool:
         """
         GOAL
         ----
         checks if SOURCE is iterable, but not exactly str!!!
         """
-        return TypeChecker.check__iterable(source, str_and_bytes_as_iterable=False)
+        return self.check__iterable(str_and_bytes_as_iterable=False)
 
     # CALLABLE --------------------------------------------------------------------------------------------------------
-    @staticmethod
-    def check__callable_func_meth_inst_cls(source: Any) -> bool:
+    def check__callable_func_meth_inst_cls(self) -> bool:
         """
         GOAL
         ----
@@ -155,10 +146,9 @@ class TypeChecker:
         ---------------------
         just to see difference and clearly using by name!!!
         """
-        return callable(source)
+        return callable(self.SOURCE)
 
-    @staticmethod
-    def check__callable_func_meth_inst(source: Any) -> bool:
+    def check__callable_func_meth_inst(self) -> bool:
         """
         GOAL
         ----
@@ -170,18 +160,16 @@ class TypeChecker:
         detect all funcs like func/meth/or even DescriptedClasses (it is class but actually used like func!)
         recommended using instead of just Callable! cause Callable keeps additionally every class instead of just simple func/method!
         """
-        if TypeChecker.check__class(source):
-            result = issubclass(source, TYPES.ELEMENTARY)
+        if self.check__class():
+            result = issubclass(self.SOURCE, TYPES.ELEMENTARY)
         else:
-            result = callable(source)
+            result = callable(self.SOURCE)
         return result
 
-    @staticmethod
-    def check__callable_func_meth(source: Any) -> bool:
-        return TypeChecker.check__callable_func(source) or TypeChecker.check__callable_meth(source)
+    def check__callable_func_meth(self) -> bool:
+        return self.check__callable_func() or self.check__callable_meth()
 
-    @staticmethod
-    def check__callable_func(source: Any) -> bool:
+    def check__callable_func(self) -> bool:
         """
         if only the exact generic function! no class method!
         Lambda included!
@@ -191,14 +179,13 @@ class TypeChecker:
         ---------------------
         not special! just as ones found ability to!
         """
-        if TypeChecker.check__callable_cls_as_func_builtin(source):
+        if self.check__callable_cls_as_func_builtin():
             result = True
         else:
-            result = TYPES.FUNCTION in source.__class__.__mro__
+            result = TYPES.FUNCTION in self.SOURCE.__class__.__mro__
         return result
 
-    @staticmethod
-    def check__callable_meth(source: Any) -> bool:
+    def check__callable_meth(self) -> bool:
         """
         if only the exact instance method!
         no generic funcs!
@@ -209,84 +196,65 @@ class TypeChecker:
         ---------------------
         not special! just as ones found ability to!
         """
-        result = not TypeChecker.check__class(source) and type(_Cls().meth) in source.__class__.__mro__
+        result = not self.check__class() and type(_Cls().meth) in self.SOURCE.__class__.__mro__
         return result
 
-    @staticmethod
-    def check__callable_inst(source: Any) -> bool:
+    def check__callable_inst(self) -> bool:
         """
         CREATED SPECIALLY FOR
         ---------------------
         not special! just as ones found ability to!
         """
-        result = TypeChecker.check__instance(source) and hasattr(source, "__call__")
+        result = self.check__instance() and hasattr(self.SOURCE, "__call__")
         return result
 
-    @staticmethod
-    def check__callable_cls_as_func_builtin(source: Any) -> bool:
+    def check__callable_cls_as_func_builtin(self) -> bool:
         """
         if class and class is as func like int/str/*  or nested
         """
-        return TypeChecker.check__class(source) and issubclass(source, TYPES.ELEMENTARY)
+        return self.check__class() and issubclass(self.SOURCE, TYPES.ELEMENTARY)
 
     # CLS/INST --------------------------------------------------------------------------------------------------------
-    @staticmethod
-    def check__class(source: Any) -> bool:
+    def check__class(self) -> bool:
         """
         works both for funcs/meths for any Сды/Штые1 see tests test__check__class
         """
-        # return hasattr(source, "__class__")     # this is incorrect!!! tests get fail!
+        # return hasattr(self.SOURCE, "__class__")     # this is incorrect!!! tests get fail!
         try:
-            return issubclass(source, object)
+            return issubclass(self.SOURCE, object)
         except:
             return False
 
-    @staticmethod
-    def check__instance(source: Any) -> bool:
-        return not TypeChecker.check__class(source) and not TypeChecker.check__callable_func(source) and not TypeChecker.check__callable_meth(source)
+    def check__instance(self) -> bool:
+        return not self.check__class() and not self.check__callable_func() and not self.check__callable_meth()
 
-    @staticmethod
-    def check__instance_not_elementary(source: Any) -> bool:
-        return TypeChecker.check__instance(source) and not TypeChecker.check__elementary(source)
+    def check__instance_not_elementary(self) -> bool:
+        return self.check__instance() and not self.check__elementary()
 
-    @staticmethod
-    def check__exception(source: Any) -> bool:
+    def check__exception(self) -> bool:
         """
         any of both variant (Instance/Class) of any Exception!
         """
-        if isinstance(source, Exception):
+        if isinstance(self.SOURCE, Exception):
             return True
         try:
-            return issubclass(source, Exception)
+            return issubclass(self.SOURCE, Exception)
         except:
             pass
         return False
 
-    @staticmethod
-    def check__nested__by_cls_or_inst(source: Any, parent: Any) -> bool | None:
+    def check__nested__by_cls_or_inst(self, parent: Any) -> bool | None:
         """
         any of both variant (Instance/Class) comparing with TARGET of both variant (Instance/Class)
 
         specially created for pytest_aux for comparing with Exception!
         """
-        # source = Lambda(source).get_result_or_exx()
-        # parent = Lambda(parent).get_result_or_exx()
+        source_cls = self.ensure__class()
+        parent_cls = TypeCheck(parent).ensure__class()
+        return issubclass(source_cls, parent_cls)
 
-        if TypeChecker.check__instance(source):
-            source = source.__class__
-        if TypeChecker.check__instance(parent):
-            parent = parent.__class__
-
-        if not TypeChecker.check__class(source) or not TypeChecker.check__class(parent):
-            return None
-
-        return issubclass(source, parent)
-
-
-# =====================================================================================================================
-class TypeEnsure:
-    @staticmethod
-    def ensure__class(source: Any) -> type:
+    # =================================================================================================================
+    def ensure__class(self) -> type:
         """
         GOAL
         ----
@@ -296,10 +264,10 @@ class TypeEnsure:
         ---------------------
         classes.ClsMiddleGroup
         """
-        if TypeChecker.check__class(source):
-            return source
+        if self.check__class():
+            return self.SOURCE
         else:
-            return source.__class__
+            return self.SOURCE.__class__
 
 
 # =====================================================================================================================