+++ tests/objects/test__types.py
@@ -2,7 +2,6 @@ from typing import *
 import pytest
 
 from base_aux.base_objects import *
-
 from base_aux.pytester import *
 
 
@@ -13,24 +12,9 @@ CLASSES__AS_FUNC: list = [ClsInt, ClsStr, ClsList, ClsSet, ClsDict, ]   # actual
 
 # =====================================================================================================================
 class Test__1:
-    @classmethod
-    def setup_class(cls):
-        pass
-        cls.Victim = TypeChecker
-
-    # @classmethod
-    # def teardown_class(cls):
-    #     pass
-    #
-    # def setup_method(self, method):
-    #     pass
-    #
-    # def teardown_method(self, method):
-    #     pass
-
     # -----------------------------------------------------------------------------------------------------------------
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             ("_", False),
             ("__", False),
@@ -50,207 +34,76 @@ class Test__1:
             ("abc", False),
         ]
     )
-    def test__name_is_build_in(self, args, _EXPECTED):
-        victim = self.Victim.check__name_is_build_in
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__name_is_build_in(self, source, _EXPECTED):
+        func_link = TypeCheck._name_is_buildin(source)
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     # =================================================================================================================
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
-            (None, True),
-            (True, True),
-            (False, True),
-            (0, False),
-            (111, False),
-            (111.222, False),
-            ("str", False),
-            (b"bytes", False),
+            (None,      (True, True, True, False, False, )),
+            (True,      (True, True, True, True, False, )),
+            (False,     (True, True, True, True, False, )),
+            (0,         (False, True, True, True, False, )),
+            (111,       (False, True, True, True, False, )),
+            (111.222,   (False, True, True, True, False, )),
+            ("str",     (False, True, True, True, False, )),
+            (b"bytes",  (False, True, True, True, False, )),
 
-            (((111, ),), False),
-            (([111, ],), False),
-            (({111, },), False),
-            (({111: 222, },), False),
+            ((111, ),        (False, True, False, False, True, )),
+            ([111, ],        (False, True, False, False, True, )),
+            ({111, },        (False, True, False, False, True, )),
+            ({111: 222, },   (False, True, False, False, True, )),
 
-            (int, False),
-            (int(1), False),
-            (str, False),
-            (str(1), False),
+            (int,       (False, False, False, False, False, )),
+            (int(1),    (False, True, True, True, False, )),
+            (str,       (False, False, False, False, False, )),
+            (str(1),    (False, True, True, True, False, )),
 
-            (Exception, False),
-            (Exception(), False),
-            (ClsException, False),
-            (ClsException(), False),
+            (Exception,     (False, False, False, False, False, )),
+            (Exception(),   (False, False, False, False, False, )),
+            (ClsException,  (False, False, False, False, False, )),
+            (ClsException(), (False, False, False, False,False,  )),
 
-            (Cls, False),
-            (Cls(), False),
-            (ClsInt, False),
-            (ClsInt(), False),    # int() == 0!!!
+            (Cls,       (False, False, False, False, False, )),
+            (Cls(),     (False, False, False, False, False, )),
+            (ClsInt,    (False, False, False, False, False, )),
+            (ClsInt(),  (False, True, True, True, False, )),    # int() == 0!!!
 
-            (FUNC, False),
-            (LAMBDA, False),
-            (ClsCallNone, False),
-            (ClsCallNone(), False),
-            (ClsCallNone()(), True),
-            (ClsCall().meth, False),
-            (ClsFullTypes.attrNone, True),
-            (ClsFullTypes().attrNone, True),
+            (FUNC,                      (False, False, False, False, False, )),
+            (LAMBDA,                    (False, False, False, False, False, )),
+            (ClsCallNone,               (False, False, False, False, False, )),
+            (ClsCallNone(),             (False, False, False, False, False, )),
+            (ClsCallNone()(),           (True, True, True, False, False, )),
+            (ClsCall().meth,            (False, False, False, False, False, )),
+            (ClsFullTypes.attrNone,     (True, True, True, False, False, )),
+            (ClsFullTypes().attrNone,   (True, True, True, False, False, )),
 
-            *[(class_i, False) for class_i in CLASSES__AS_FUNC]
+            *[
+                (
+                    class_i,
+                    (False, False, False, False, False, )
+                ) for class_i in CLASSES__AS_FUNC
+            ]
         ]
     )
-    def test__check__bool_none(self, args, _EXPECTED):
-        victim = self.Victim.check__bool_none
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__bool_none(self, source, _EXPECTED):
+        victim = TypeCheck(source)
+
+        pytest_func_tester__no_args_kwargs(victim.check__bool_none, _EXPECTED[0])
+        pytest_func_tester__no_args_kwargs(victim.check__elementary, _EXPECTED[1])
+        pytest_func_tester__no_args_kwargs(victim.check__elementary_single, _EXPECTED[2])
+        pytest_func_tester__no_args_kwargs(victim.check__elementary_single_not_none, _EXPECTED[3])
+        pytest_func_tester__no_args_kwargs(victim.check__elementary_collection, _EXPECTED[4])
+
+
+
+
+
 
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
-        argvalues=[
-            (None, True),
-            (True, True),
-            (False, True),
-            (0, True),
-            (111, True),
-            (111.222, True),
-            ("str", True),
-            (b"bytes", True),
-
-            (((111, ),), True),
-            (([111, ],), True),
-            (({111, },), True),
-            (({111: 222, },), True),
-
-            (int, False),
-            (int(1), True),
-            (str, False),
-            (str(1), True),
-
-            (Exception, False),
-            (Exception(), False),
-            (ClsException, False),
-            (ClsException(), False),
-
-            (Cls, False),
-            (Cls(), False),
-            (ClsInt, False),
-            (ClsInt(), True),    # int() == 0!!!
-
-            (FUNC, False),
-            (LAMBDA, False),
-            (ClsCallNone, False),
-            (ClsCallNone(), False),
-            (ClsCallNone()(), True),
-            (ClsCall().meth, False),
-            (ClsFullTypes.attrNone, True),
-            (ClsFullTypes().attrNone, True),
-
-            *[(class_i, False) for class_i in CLASSES__AS_FUNC]
-        ]
-    )
-    def test__check__elementary(self, args, _EXPECTED):
-        victim = self.Victim.check__elementary
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
-
-    @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
-        argvalues=[
-            (None, True),
-            (True, True),
-            (False, True),
-            (0, True),
-            (111, True),
-            (111.222, True),
-            ("str", True),
-            (b"bytes", True),
-
-            (((111, ),), False),
-            (([111, ],), False),
-            (({111, },), False),
-            (({111: 222, },), False),
-
-            (int, False),
-            (int(1), True),
-            (str, False),
-            (str(1), True),
-
-            (Exception, False),
-            (Exception(), False),
-            (ClsException, False),
-            (ClsException(), False),
-
-            (Cls, False),
-            (Cls(), False),
-            (ClsInt, False),
-            (ClsInt(), True),    # int() == 0!!!
-
-            (FUNC, False),
-            (LAMBDA, False),
-            (ClsCallNone, False),
-            (ClsCallNone(), False),
-            (ClsCallNone()(), True),
-            (ClsCall.meth, False),
-            (ClsCall().meth, False),
-            (ClsFullTypes.attrNone, True),
-            (ClsFullTypes().attrNone, True),
-
-            *[(class_i, False) for class_i in CLASSES__AS_FUNC]
-        ]
-    )
-    def test__check__elementary_single(self, args, _EXPECTED):
-        victim = self.Victim.check__elementary_single
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
-
-    @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
-        argvalues=[
-            (None, False),
-            (True, True),
-            (False, True),
-            (0, True),
-            (111, True),
-            (111.222, True),
-            ("str", True),
-            (b"bytes", True),
-
-            (((111, ),), False),
-            (([111, ],), False),
-            (({111, },), False),
-            (({111: 222, },), False),
-
-            (int, False),
-            (int(1), True),
-            (str, False),
-            (str(1), True),
-
-            (Exception, False),
-            (Exception(), False),
-            (ClsException, False),
-            (ClsException(), False),
-
-            (Cls, False),
-            (Cls(), False),
-            (ClsInt, False),
-            (ClsInt(), True),    # int() == 0!!!
-
-            (FUNC, False),
-            (LAMBDA, False),
-            (ClsCallNone, False),
-            (ClsCallNone(), False),
-            (ClsCallNone()(), False),
-            (ClsCall.meth, False),
-            (ClsCall().meth, False),
-            (ClsFullTypes.attrNone, False),
-            (ClsFullTypes().attrNone, False),
-
-            *[(class_i, False) for class_i in CLASSES__AS_FUNC]
-        ]
-    )
-    def test__check__elementary_single_not_none(self, args, _EXPECTED):
-        victim = self.Victim.check__elementary_single_not_none
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
-
-    @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             (None, False),
             (True, False),
@@ -261,10 +114,10 @@ class Test__1:
             ("str", False),
             (b"bytes", False),
 
-            (((111, ),), True),
-            (([111, ],), True),
-            (({111, },), True),
-            (({111: 222, },), True),
+            ((111, ), True),
+            ([111, ], True),
+            ({111, }, True),
+            ({111: 222, }, False),
 
             (int, False),
             (int(1), False),
@@ -294,120 +147,71 @@ class Test__1:
             *[(class_i, False) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__check__elementary_collection(self, args, _EXPECTED):
-        victim = self.Victim.check__elementary_collection
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
-
-    @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
-        argvalues=[
-            (None, False),
-            (True, False),
-            (False, False),
-            (0, False),
-            (111, False),
-            (111.222, False),
-            ("str", False),
-            (b"bytes", False),
-
-            (((111, ),), True),
-            (([111, ],), True),
-            (({111, },), True),
-            (({111: 222, },), False),
-
-            (int, False),
-            (int(1), False),
-            (str, False),
-            (str(1), False),
-
-            (Exception, False),
-            (Exception(), False),
-            (ClsException, False),
-            (ClsException(), False),
-
-            (Cls, False),
-            (Cls(), False),
-            (ClsInt, False),
-            (ClsInt(), False),    # int() == 0!!!
-
-            (FUNC, False),
-            (LAMBDA, False),
-            (ClsCallNone, False),
-            (ClsCallNone(), False),
-            (ClsCallNone()(), False),
-            (ClsCall.meth, False),
-            (ClsCall().meth, False),
-            (ClsFullTypes.attrNone, False),
-            (ClsFullTypes().attrNone, False),
-
-            *[(class_i, False) for class_i in CLASSES__AS_FUNC]
-        ]
-    )
-    def test__check__elementary_collection_not_dict(self, args, _EXPECTED):
-        victim = self.Victim.check__elementary_collection_not_dict
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__elementary_collection_not_dict(self, source, _EXPECTED):
+        func_link = TypeCheck(source).check__elementary_collection_not_dict
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     # -----------------------------------------------------------------------------------------------------------------
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, args, _EXPECTED",
         argvalues=[
-            (("str", True, True), True),
-            (("str", True, False), False),
+            ("str", (True, True), True),
+            ("str", (True, False), False),
 
-            ((b"bytes", True, True), True),
-            ((b"bytes", True, False), False),
+            (b"bytes", (True, True), True),
+            (b"bytes", (True, False), False),
 
             # -----------------------
-            (None, False),
-            (True, False),
-            (False, False),
-            (0, False),
-            (111, False),
-            (111.222, False),
-            ("str", True),
-            (b"bytes", True),
+            (None, (), False),
+            (True, (), False),
+            (False, (), False),
+            (0, (), False),
+            (111, (), False),
+            (111.222, (), False),
+            ("str", (), True),
+            (b"bytes", (), True),
 
-            (((111, ),), True),
-            (([111, ],), True),
-            (({111, },), True),
-            (({111: 222, },), True),
-            (({111: 222, }, True, True), True),
-            (({111: 222, }, False, True), False),
+            ((111, ), (), True),
+            ([111, ], (), True),
+            ({111, }, (), True),
+            ({111: 222, }, (), True),
+            ({111: 222, }, (True, True), True),
+            ({111: 222, }, (False, True), False),
 
-            (int, False),
-            (int(1), False),
-            (str, True),        # not clear!!!
-            (str(1), True),
+            (int, (), False),
+            (int(1), (), False),
+            (str, (), True),        # not clear!!!
+            (str(1), (), True),
 
-            (Exception, False),
-            (Exception(), False),
-            (ClsException, False),
-            (ClsException(), False),
+            (Exception, (), False),
+            (Exception(), (), False),
+            (ClsException, (), False),
+            (ClsException(), (), False),
 
-            (Cls, False),
-            (Cls(), False),
-            (ClsInt, False),
-            (ClsInt(), False),
+            (Cls, (), False),
+            (Cls(), (), False),
+            (ClsInt, (), False),
+            (ClsInt(), (), False),
 
-            (FUNC, False),
-            (LAMBDA, False),
-            (ClsCallNone, False),
-            (ClsCallNone(), False),
-            (ClsCallNone()(), False),
-            (ClsCall.meth, False),
-            (ClsCall().meth, False),
-            (ClsFullTypes.attrNone, False),
-            (ClsFullTypes().attrNone, False),
+            (FUNC, (), False),
+            (LAMBDA, (), False),
+            (ClsCallNone, (), False),
+            (ClsCallNone(),(),  False),
+            (ClsCallNone()(), (), False),
+            (ClsCall.meth, (), False),
+            (ClsCall().meth, (), False),
+            (ClsFullTypes.attrNone, (), False),
+            (ClsFullTypes().attrNone, (), False),
 
             # *[(class_i, False) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__iterable(self, args, _EXPECTED):
-        victim = self.Victim.check__iterable
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__iterable(self, source, args, _EXPECTED):
+        func_link = TypeCheck(source).check__iterable
+        pytest_func_tester__no_kwargs(func_link, args, _EXPECTED)
 
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             (None, False),
             (True, False),
@@ -451,13 +255,13 @@ class Test__1:
             # *[(class_i, False) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__iterable_not_str(self, args, _EXPECTED):
-        victim = self.Victim.check__iterable_not_str
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__iterable_not_str(self, source, _EXPECTED):
+        func_link = TypeCheck(source).check__iterable_not_str
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     # CALLABLE --------------------------------------------------------------------------------------------------------
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             (None, False),
             (True, False),
@@ -501,12 +305,12 @@ class Test__1:
             *[(class_i, True) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__check__callable_func_meth_inst(self, args, _EXPECTED):
-        victim = self.Victim.check__callable_func_meth_inst
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__callable_func_meth_inst(self, source, _EXPECTED):
+        func_link = TypeCheck(source).check__callable_func_meth_inst
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             (None, False),
             (True, False),
@@ -550,12 +354,12 @@ class Test__1:
             *[(class_i, True) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__check__callable_func_meth(self, args, _EXPECTED):
-        victim = self.Victim.check__callable_func_meth
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__callable_func_meth(self, source, _EXPECTED):
+        func_link = TypeCheck(source).check__callable_func_meth
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             (None, False),
             (True, False),
@@ -599,12 +403,12 @@ class Test__1:
             *[(class_i, True) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__check__callable_func(self, args, _EXPECTED):
-        victim = self.Victim.check__callable_func
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__callable_func(self, source, _EXPECTED):
+        func_link = TypeCheck(source).check__callable_func
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             (None, False),
             (True, False),
@@ -648,12 +452,12 @@ class Test__1:
             *[(class_i, False) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__check__callable_meth(self, args, _EXPECTED):
-        victim = self.Victim.check__callable_meth
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__callable_meth(self, source, _EXPECTED):
+        func_link = TypeCheck(source).check__callable_meth
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             (None, False),
             (True, False),
@@ -697,12 +501,12 @@ class Test__1:
             *[(class_i, False) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__check__callable_inst(self, args, _EXPECTED):
-        victim = self.Victim.check__callable_inst
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__callable_inst(self, source, _EXPECTED):
+        func_link = TypeCheck(source).check__callable_inst
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             (None, False),
             (True, False),
@@ -746,13 +550,13 @@ class Test__1:
             *[(class_i, True) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__check__callable_cls_as_func_builtin(self, args, _EXPECTED):
-        victim = self.Victim.check__callable_cls_as_func_builtin
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__callable_cls_as_func_builtin(self, source, _EXPECTED):
+        func_link = TypeCheck(source).check__callable_cls_as_func_builtin
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     # -----------------------------------------------------------------------------------------------------------------
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             (None, False),
             (True, False),
@@ -796,12 +600,12 @@ class Test__1:
             *[(class_i, True) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__check__class(self, args, _EXPECTED):
-        victim = self.Victim.check__class
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__class(self, source, _EXPECTED):
+        func_link = TypeCheck(source).check__class
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             (None, True),
             (True, True),
@@ -845,12 +649,12 @@ class Test__1:
             *[(class_i, False) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__check__instance(self, args, _EXPECTED):
-        victim = self.Victim.check__instance
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__instance(self, source, _EXPECTED):
+        func_link = TypeCheck(source).check__instance
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             (None, False),
             (True, False),
@@ -894,13 +698,13 @@ class Test__1:
             *[(class_i, False) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__check__instance_not_elementary(self, args, _EXPECTED):
-        victim = self.Victim.check__instance_not_elementary
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__instance_not_elementary(self, source, _EXPECTED):
+        func_link = TypeCheck(source).check__instance_not_elementary
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     # -----------------------------------------------------------------------------------------------------------------
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, _EXPECTED",
         argvalues=[
             (None, False),
             (True, False),
@@ -944,54 +748,54 @@ class Test__1:
             *[(class_i, False) for class_i in CLASSES__AS_FUNC]
         ]
     )
-    def test__check__exception(self, args, _EXPECTED):
-        victim = self.Victim.check__exception
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__exception(self, source, _EXPECTED):
+        func_link = TypeCheck(source).check__exception
+        pytest_func_tester__no_args_kwargs(func_link, _EXPECTED)
 
     # =================================================================================================================
     @pytest.mark.parametrize(
-        argnames="args, _EXPECTED",
+        argnames="source, parent, _EXPECTED",
         argvalues=[
-            (("str", "str"), True),
-            (("str", str), True),
-            ((str, "str"), True),
-            ((str, str), True),
+            ("str", "str", True),
+            ("str", str, True),
+            (str, "str", True),
+            (str, str, True),
 
-            ((int, str), False),
-            ((int, "str"), False),
+            (int, str, False),
+            (int, "str", False),
 
-            ((111, 111), True),
-            ((int, 111), True),
-            ((111, int), True),
-            ((int, int), True),
+            (111, 111, True),
+            (int, 111, True),
+            (111, int, True),
+            (int, int, True),
 
-            ((Exception, Exception), True),
-            ((Exception(), Exception), True),
-            ((Exception, Exception()), True),
-            ((Exception(), Exception()), True),
+            (Exception, Exception, True),
+            (Exception(), Exception, True),
+            (Exception, Exception(), True),
+            (Exception(), Exception(), True),
 
-            ((ClsException, Exception), True),
-            ((ClsException(), Exception), True),
-            ((ClsException, Exception()), True),
-            ((ClsException(), Exception()), True),
+            (ClsException, Exception, True),
+            (ClsException(), Exception, True),
+            (ClsException, Exception(), True),
+            (ClsException(), Exception(), True),
 
-            ((Exception, ClsException), False),      # REMEMBER! not clear!
-            ((Exception(), ClsException), False),    # REMEMBER! not clear!
-            ((Exception, ClsException()), False),    # REMEMBER! not clear!
-            ((Exception(), ClsException()), False),  # REMEMBER! not clear!
+            (Exception, ClsException, False),      # REMEMBER! not clear!
+            (Exception(), ClsException, False),    # REMEMBER! not clear!
+            (Exception, ClsException(), False),    # REMEMBER! not clear!
+            (Exception(), ClsException(), False),  # REMEMBER! not clear!
 
-            ((Cls, Cls), True),
-            ((Cls, Cls()), True),
-            ((Cls(), Cls), True),
-            ((Cls(), Cls()), True),
+            (Cls, Cls, True),
+            (Cls, Cls(), True),
+            (Cls(), Cls, True),
+            (Cls(), Cls(), True),
 
-            ((FUNC, Cls), None),
-            ((FUNC, Cls()), None),
+            (FUNC, Cls, False),
+            (FUNC, Cls(), False),
         ]
     )
-    def test__check__nested__by_cls_or_inst(self, args, _EXPECTED):
-        victim = self.Victim.check__nested__by_cls_or_inst
-        pytest_func_tester__no_kwargs(victim, args, _EXPECTED)
+    def test__check__nested__by_cls_or_inst(self, source, parent, _EXPECTED):
+        func_link = TypeCheck(source).check__nested__by_cls_or_inst
+        pytest_func_tester__no_kwargs(func_link, parent, _EXPECTED)
 
 
 # =====================================================================================================================