+++ HISTORY.md
@@ -18,7 +18,7 @@
 
 0.1.8 (2024/12/11 11:41:19)
 ------------------------------
-- start separating parts [exceptions/TypeEnsure/pytester/attrs]   
+- start separating parts [exceptions/TypeCheck/pytester/attrs]   
 
 0.1.7 (2024/12/09 17:03:42)
 ------------------------------